requirements_test.txt
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-homeassistant-custom-component>=0.13.0
pylint>=3.0.0
dnspython>=2.4.0
pytest-mock>=3.12.0 